chore: add bottom and top nav bar ui
