+++ lib/components/navigation_bar/navigation_bar.dart
@@ -1,0 +1,39 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+import 'navigation_bar_cubit.dart';
+
+const _navigationBarHeight = 96.0;
+
+class HotelynNavigationBar extends StatelessWidget {
+  const HotelynNavigationBar({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final cubit = context.watch<NavigationBarCubit>();
+    return BottomNavigationBar(
+      onTap: (index) => cubit.updateSelectedIndex(index),
+      currentIndex: cubit.state.selectedTabIndex,
+      showSelectedLabels: true,
+      type: BottomNavigationBarType.fixed,
+      items: const [
+        BottomNavigationBarItem(
+          icon: Icon(Icons.home),
+          label: 'Home',
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.search),
+          label: 'Search',
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.message),
+          label: 'Messages',
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.person_outline),
+          label: 'Profile',
+        )
+      ],
+    );
+  }
+}
