+++ lib/components/navigation_bar/navigation_bar_state.dart
@@ -1,0 +1,13 @@
+class NavigationBarState {
+  final int selectedTabIndex;
+
+  NavigationBarState({
+    required this.selectedTabIndex,
+  });
+
+  NavigationBarState copyWith({int? selectedTabIndex}) {
+    return NavigationBarState(
+      selectedTabIndex: selectedTabIndex ?? this.selectedTabIndex,
+    );
+  }
+}
