+++ lib/features/home/home_cubit.dart
@@ -1,0 +1,6 @@
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:hotelyn/features/home/home_state.dart';
+
+class HomeCubit extends Cubit<HomeState> {
+  HomeCubit(super.initialState);
+}
