+++ lib/features/home/home_state.dart
@@ -1,0 +1,1 @@
+sealed class HomeState {}
