+++ lib/features/messages/messages_tab.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class MesssagesTab extends StatelessWidget {
+  const MesssagesTab({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
