+++ lib/features/onboarding/on_boarding_welcome_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:hotelyn/components/buttons/hotelyn_button.dart';
 import 'package:hotelyn/components/icons/hotelyn_icon.dart';
 import 'package:hotelyn/components/text_style/hotelyn_text_style.dart';
-import 'package:hotelyn/features/home/home_page.dart';
+import 'package:hotelyn/features/home/home_tab.dart';
 
 class OnBoardingWelcomePage extends StatelessWidget {
   const OnBoardingWelcomePage({super.key});
