+++ lib/features/profile/profile_tab.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class ProfileTab extends StatelessWidget {
+  const ProfileTab({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
