+++ lib/features/search/search_tab.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class SearchTab extends StatelessWidget {
+  const SearchTab({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
